Consolidate decorators from `method` & `property` into `field`
