+++ lib/decorators/index.ts
@@ -1,2 +1,1 @@
-export * from "./method";
-export * from "./property";
+export * from "./field";
